modificacion tipo de datos en campos en diferentes de modelos

+++ api_prenar/models/categoria_material.py
@@ -4,4 +4,4 @@ class CategoriaMaterial(models.Model):
     id=models.AutoField(primary_key=True)
     name=models.CharField(max_length=255)
     color=models.CharField(max_length=255, null=True, blank=True)
-    stock_quantity=models.IntegerField(default=0)
+    stock_quantity=models.FloatField(default=0.0)
